Refactor ProjectsGrid into a shared component. refactor homepage and projects to use their own query and use the ProjectsGrid to display the information
