+++ client/src/common/components/ProjectsGrid/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './ProjectsGrid';
