+++ client/src/pages/Projects/GetProjectsQuery.tsx
@@ -10,7 +10,6 @@ const PROJECT_QUERY = gql`
       projects {
         id
         title
-        meaningToTheKids
       }
     }
   }
